Move new queue to main queue and update references
