--- queue/new/queue.css
@@ -1,65 +1,0 @@
-.container {
-	display: table-cell;
-	background-color: #ebebea;
-}
-
-.container div,
-.gu-mirror {
-	margin: 10px 5px;
-	padding: 10px;
-	background-color: rgba(0, 0, 0, 0.2);
-	transition: opacity 0.4s ease-in-out;
-	border-radius: 8px;
-  	box-shadow: rgba(0, 0, 0, 0.5) 0px 2px 2px -2px, rgba(0, 0, 0, 0.2) 0px 1px 3px;
-}
-
-.container div:hover, .gu-mirror:hover {
-    background-color: rgba(0, 0, 0, 0.1);
-}
-
-.container div {
-	cursor: move;
-	cursor: grab;
-	cursor: -moz-grab;
-	cursor: -webkit-grab;
-}
-
-.gu-mirror {
-	cursor: grabbing;
-	cursor: -moz-grabbing;
-	cursor: -webkit-grabbing;
-}
-
-.pill {
-	display: inline-block;
-	padding: .25em .6em;
-	line-height: 1;
-	text-align: center;
-	white-space: nowrap;
-	vertical-align: baseline;
-	border-radius: 10rem;
-}
-
-.pill-complete {
-	color: #fff;
-	background-color: #137cbd;
-}
-
-.pill-incomplete {
-	border: 1px solid grey;
-}
-
-.pill-red {
-	color: #fff;
-	background-color: #D00000;
-}
-
-.pill-green {
-	color: #fff;
-	background-color: #38764C;
-}
-
-.pill-walk-in {
-	color: #fff;
-	background-color: #38764C;
-}
