--- queue/queueSearchFilter.js
@@ -1,37 +1,0 @@
-define('queueSearchFilter', [], function() {
-
-	function searchFor() {
-
-		// All filters must return a function. The first parameter
-		// is the data that is to be filtered, and the second is an
-		// argument that may be passed with a colon (searchFor:searchString)
-
-		return function(arr, searchString) {
-
-			if(!searchString) {
-				return arr;
-			}
-
-			var result = [];
-
-			searchString = searchString.toLowerCase();
-
-			// Using the forEach helper method to loop through the array
-			angular.forEach(arr, function(item) {
-
-				if(item.name.toLowerCase().indexOf(searchString) !== -1 ||
-					item.appointmentId.toString().indexOf(searchString) !== -1 ||
-					("#" + item.appointmentId.toString()).indexOf(searchString) !== -1){
-					result.push(item);
-				}
-
-			});
-
-			return result;
-		};
-
-	}
-
-	return searchFor;
-
-});
